+++ scanpy/neighbors/__init__.py
@@ -8,17 +8,20 @@ from typing import (
     NamedTuple,
     Tuple,
     Literal,
+    MutableMapping,
+    Mapping,
+    Callable,
 )
-from collections.abc import MutableMapping, Mapping, Callable, Generator
-import warnings
 
 import numpy as np
 import scipy
 from anndata import AnnData
-from scipy.sparse import issparse, coo_matrix, csr_matrix
+from scipy.sparse import issparse, csr_matrix
 from sklearn.utils import check_random_state
 
 from .enums import _Metric, _MetricFn, _Method
+from .backends.umap import compute_neighbors_umap, _compute_connectivities_umap
+from .backends.rapids import compute_neighbors_rapids
 from .. import logging as logg
 from .. import _utils
 from .._utils import _doc_params, AnyRandom, NeighborsView
@@ -179,222 +182,6 @@ class FlatTree(NamedTuple):
 
 
 RPForestDict = Mapping[str, Mapping[str, np.ndarray]]
-
-
-def _rp_forest_generate(
-    rp_forest_dict: RPForestDict,
-) -> Generator[FlatTree, None, None]:
-    props = FlatTree._fields
-    num_trees = len(rp_forest_dict[props[0]]['start']) - 1
-
-    for i in range(num_trees):
-        tree = []
-        for prop in props:
-            start = rp_forest_dict[prop]['start'][i]
-            end = rp_forest_dict[prop]['start'][i + 1]
-            tree.append(rp_forest_dict[prop]['data'][start:end])
-        yield FlatTree(*tree)
-
-    tree = []
-    for prop in props:
-        start = rp_forest_dict[prop]['start'][num_trees]
-        tree.append(rp_forest_dict[prop]['data'][start:])
-    yield FlatTree(*tree)
-
-
-def compute_neighbors_umap(
-    X: Union[np.ndarray, csr_matrix],
-    n_neighbors: int,
-    random_state: AnyRandom = None,
-    metric: Union[_Metric, _MetricFn] = 'euclidean',
-    metric_kwds: Mapping[str, Any] = MappingProxyType({}),
-    angular: bool = False,
-    verbose: bool = False,
-):
-    """This is from umap.fuzzy_simplicial_set [McInnes18]_.
-
-    Given a set of data X, a neighborhood size, and a measure of distance
-    compute the fuzzy simplicial set (here represented as a fuzzy graph in
-    the form of a sparse matrix) associated to the data. This is done by
-    locally approximating geodesic distance at each point, creating a fuzzy
-    simplicial set for each such point, and then combining all the local
-    fuzzy simplicial sets into a global one via a fuzzy union.
-
-    Parameters
-    ----------
-    X: array of shape (n_samples, n_features)
-        The data to be modelled as a fuzzy simplicial set.
-    n_neighbors
-        The number of neighbors to use to approximate geodesic distance.
-        Larger numbers induce more global estimates of the manifold that can
-        miss finer detail, while smaller values will focus on fine manifold
-        structure to the detriment of the larger picture.
-    random_state
-        A state capable being used as a numpy random state.
-    metric
-        The metric to use to compute distances in high dimensional space.
-        If a string is passed it must match a valid predefined metric. If
-        a general metric is required a function that takes two 1d arrays and
-        returns a float can be provided. For performance purposes it is
-        required that this be a numba jit'd function. Valid string metrics
-        include:
-            * euclidean
-            * manhattan
-            * chebyshev
-            * minkowski
-            * canberra
-            * braycurtis
-            * mahalanobis
-            * wminkowski
-            * seuclidean
-            * cosine
-            * correlation
-            * haversine
-            * hamming
-            * jaccard
-            * dice
-            * russelrao
-            * kulsinski
-            * rogerstanimoto
-            * sokalmichener
-            * sokalsneath
-            * yule
-        Metrics that take arguments (such as minkowski, mahalanobis etc.)
-        can have arguments passed via the metric_kwds dictionary. At this
-        time care must be taken and dictionary elements must be ordered
-        appropriately; this will hopefully be fixed in the future.
-    metric_kwds
-        Arguments to pass on to the metric, such as the ``p`` value for
-        Minkowski distance.
-    angular
-        Whether to use angular/cosine distance for the random projection
-        forest for seeding NN-descent to determine approximate nearest
-        neighbors.
-    verbose
-        Whether to report information on the current progress of the algorithm.
-
-    Returns
-    -------
-    **knn_indices**, **knn_dists** : np.arrays of shape (n_observations, n_neighbors)
-    """
-    with warnings.catch_warnings():
-        # umap 0.5.0
-        warnings.filterwarnings("ignore", message=r"Tensorflow not installed")
-        from umap.umap_ import nearest_neighbors
-
-    random_state = check_random_state(random_state)
-
-    knn_indices, knn_dists, forest = nearest_neighbors(
-        X,
-        n_neighbors,
-        random_state=random_state,
-        metric=metric,
-        metric_kwds=metric_kwds,
-        angular=angular,
-        verbose=verbose,
-        n_jobs=settings.n_jobs,
-    )
-
-    return knn_indices, knn_dists, forest
-
-
-def compute_neighbors_rapids(
-    X: np.ndarray, n_neighbors: int, metric: Union[_Metric, _MetricFn] = 'euclidean'
-):
-    """Compute nearest neighbors using RAPIDS cuml.
-
-    Parameters
-    ----------
-    X: array of shape (n_samples, n_features)
-        The data to compute nearest neighbors for.
-    n_neighbors
-        The number of neighbors to use.
-    metric
-        The metric to use to compute distances in high dimensional space.
-        This string must match a valid predefined metric in RAPIDS cuml.
-
-        Returns
-    -------
-    **knn_indices**, **knn_dists** : np.arrays of shape (n_observations, n_neighbors)
-    """
-    from cuml.neighbors import NearestNeighbors
-
-    nn = NearestNeighbors(n_neighbors=n_neighbors, metric=metric)
-    X_contiguous = np.ascontiguousarray(X, dtype=np.float32)
-    nn.fit(X_contiguous)
-    knn_dist, knn_indices = nn.kneighbors(X_contiguous)
-    return knn_indices, knn_dist
-
-
-def _get_sparse_matrix_from_indices_distances_umap(
-    knn_indices, knn_dists, n_obs, n_neighbors
-):
-    rows = np.zeros((n_obs * n_neighbors), dtype=np.int64)
-    cols = np.zeros((n_obs * n_neighbors), dtype=np.int64)
-    vals = np.zeros((n_obs * n_neighbors), dtype=np.float64)
-
-    for i in range(knn_indices.shape[0]):
-        for j in range(n_neighbors):
-            if knn_indices[i, j] == -1:
-                continue  # We didn't get the full knn for i
-            if knn_indices[i, j] == i:
-                val = 0.0
-            else:
-                val = knn_dists[i, j]
-
-            rows[i * n_neighbors + j] = i
-            cols[i * n_neighbors + j] = knn_indices[i, j]
-            vals[i * n_neighbors + j] = val
-
-    result = coo_matrix((vals, (rows, cols)), shape=(n_obs, n_obs))
-    result.eliminate_zeros()
-    return result.tocsr()
-
-
-def _compute_connectivities_umap(
-    knn_indices,
-    knn_dists,
-    n_obs,
-    n_neighbors,
-    set_op_mix_ratio=1.0,
-    local_connectivity=1.0,
-):
-    """\
-    This is from umap.fuzzy_simplicial_set [McInnes18]_.
-
-    Given a set of data X, a neighborhood size, and a measure of distance
-    compute the fuzzy simplicial set (here represented as a fuzzy graph in
-    the form of a sparse matrix) associated to the data. This is done by
-    locally approximating geodesic distance at each point, creating a fuzzy
-    simplicial set for each such point, and then combining all the local
-    fuzzy simplicial sets into a global one via a fuzzy union.
-    """
-    with warnings.catch_warnings():
-        # umap 0.5.0
-        warnings.filterwarnings("ignore", message=r"Tensorflow not installed")
-        from umap.umap_ import fuzzy_simplicial_set
-
-    X = coo_matrix(([], ([], [])), shape=(n_obs, 1))
-    connectivities = fuzzy_simplicial_set(
-        X,
-        n_neighbors,
-        None,
-        None,
-        knn_indices=knn_indices,
-        knn_dists=knn_dists,
-        set_op_mix_ratio=set_op_mix_ratio,
-        local_connectivity=local_connectivity,
-    )
-
-    if isinstance(connectivities, tuple):
-        # In umap-learn 0.4, this returns (result, sigmas, rhos)
-        connectivities = connectivities[0]
-
-    distances = _get_sparse_matrix_from_indices_distances_umap(
-        knn_indices, knn_dists, n_obs, n_neighbors
-    )
-
-    return distances, connectivities.tocsr()
 
 
 def _get_sparse_matrix_from_indices_distances_numpy(
